hackproject/code/api/requirements.txt
pytest==7.2.1
fastapi==0.94.1
uvicorn==0.20.0
llama-index==0.5.25
pydantic==1.10.4
pypdf2==3.0.1
python-dotenv<=1.0.0
pytest-mock==3.10.0
googletrans==4.0.0rc1
pyTelegramBOTAPI==4.11.0
aiohttp==3.8.4
httpcore==0.9.1
httpx==0.13.3